Created new POST /merge handler which receives form data
When you visit /merge in a browser, you're met with a page that asks you
to specify what time range you'd like to merge points on local storage
from. This handler receievs the data that users enter in that form, and
will pass it along to the place where the actual point-merging logic
happens in the future.