Laid down foundation for controlling how points are merged
The merge package is very similar to the track package.

The Merger object is responsible for fetching points stored on a local
instance of the server that were created within a specific time frame,
as well as regulating and facilitating sending those points to the
remote server.

merge_info_config.json stores info about the state of the current upload
operation.