Added temporary feedback from LocalMergeHandler
In the future, this handler should respond with a 2xx response code
(I'm thinking a 204), and the web page served to the client should
change according to that status code.

If this handler responds with a 500 status code, then the web page
should display that error message; if this handler responds with a 2xx
(204) status code, then the web page should display some kind of success
screen with a button to return to the page with the form.